Unautorized request does not return CORS headers
When testing your code, everything seems to be fine, and I really like your solution over the more cumbersome alternative solutions, but there seems to be a problem when using CORS in case of an unauthorized request (401), possibly even all failed requests (4xx and 5xx):

When a request is made to a controller action decorated with [Authorize] and using an invalid or expired token, a 401 response is returned as expected, but the CORS headers are not being returned. This causes the client not being able to recognize the status code.